fix: getSmoothedSpeedKmh返回Float转Double

+++ app/src/main/java/com/chasegame/lastfreedom/MainActivity.kt
@@ -233,7 +233,7 @@ class MainActivity : AppCompatActivity() {
         }
 
         // === 速度驱动缩放 ===
-        val speedKmh = provider.getSmoothedSpeedKmh().coerceAtLeast(0.0)
+        val speedKmh = provider.getSmoothedSpeedKmh().toDouble().coerceAtLeast(0.0)
         val targetZoom = speedToZoom(speedKmh)
         // EMA 平滑，避免 zoom 跳变
         smoothedZoom = smoothedZoom + zoomAlpha * (targetZoom - smoothedZoom)
